% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GENE1_anno}
\alias{GENE1_anno}
\title{GENE1 annotation dataframe}
\format{A data frame with 17 rows and 5 variables:
\describe{
  \item{v1}{chromosome name (could either be NC^ or Chr^)}
  \item{V2}{exon start}
  \item{V3}{enxon end}
  \item{V4}{exon id}
  \item{V5}{strand}
}}
\usage{
GENE1_anno
}
\description{
GENE1 annotation dataframe
}
\keyword{datasets}
