% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{gene_anno-class}
\alias{gene_anno-class}
\title{Title Gene trascript annotation class}
\value{
gene_anno class.
}
\description{
Title Gene trascript annotation class
}
\section{Slots}{

\describe{
\item{\code{name}}{character. gene name.}

\item{\code{chr_num}}{character. the number of chromosme. eg("1", "5", "X").}

\item{\code{chr}}{character. chromosme name.}

\item{\code{chromstart}}{numeric. numeric. chromosome start based on annotaion.}

\item{\code{chromend}}{numeric. numeric. chromosome end based on annotaion.}

\item{\code{genelen}}{numeric. length of transcript.}

\item{\code{center}}{numeric. center coordinate of the transcript.}

\item{\code{anno}}{data.frame. A datafram contain all exon coordinates.}
}}

\examples{
NULL
}
