% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gginter_utils.R
\name{gene_anno}
\alias{gene_anno}
\title{Title Read in gene annotation dataframe
Gene annotation dataframe, examples can be loaded using data(GENE1_anno)}
\usage{
gene_anno(gene_anno, genename = "")
}
\arguments{
\item{gene_anno}{Dataframe, bed file format, each row represent an exon.}
}
\value{
a new gene_anno class object
}
\description{
Title Read in gene annotation dataframe
Gene annotation dataframe, examples can be loaded using data(GENE1_anno)
}
\examples{
data(GENE1_anno) anno1 <- gene_anno(GENE1_anno)
gene1 <- gene_anno(DF)
}
