% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ideo-class}
\alias{ideo-class}
\title{Title Ceate single ideogram ggplot object.}
\value{
ideo object with ggplot layer and meta information about the ideogram,
the highlight region and the plot object.
}
\description{
Title Ceate single ideogram ggplot object.
}
\section{Slots}{

\describe{
\item{\code{geom_ideobody}}{list. ggplot layer list for ideogram presentation.}

\item{\code{geom_tick}}{list. ggplot layer list for hightlight region of ideogram.}

\item{\code{chr}}{character. chromosme name of ideogram, starts with "chr".}

\item{\code{height}}{numeric. single ideogram height (horizontal layout).}

\item{\code{xdrift}}{numeric. x axis drift from 0.}

\item{\code{ydrift}}{numeric. y axis drift from 0.}

\item{\code{chr_start}}{numeric. chromosome start based on annotaion.}

\item{\code{chr_end}}{numeric. chromosome end based on annotaion.}

\item{\code{.plot_scale}}{numeric. scale compression level from real annotation to ggplot presented object.}

\item{\code{.ideo_left}}{numeric. leftmost coordinate of ideogram object (horizontal layout).}

\item{\code{.ideo_right}}{numeric. rightmost coordinate of ideogram object (horizontal layout).}

\item{\code{.tick_top}}{numeric. topmost coordinate of ideogram object (horizontal layout, if highlight region/tick added).}

\item{\code{.tick_bot}}{numeric. bottommost coordinate of ideogram object (horizontal layout, if highlight region/tick added).}

\item{\code{.leftregion}}{numeric. start of highlight region (orginal coordinate).}

\item{\code{.rightregion}}{numeric. end of highlight region (orginal coordinate).}
}}

\examples{
NULL
}
