% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideo_add_highlight.R
\name{ideo_add_highlight}
\alias{ideo_add_highlight}
\title{Title}
\usage{
ideo_add_highlight(ideo, region, tick_len = 15)
}
\arguments{
\item{ideo}{ideo object, object to add a highlight region.}

\item{region}{list, the region to highlight,}

\item{tick_len}{numeric, the vertical height of the highligh region.}
}
\value{
ideo object.
}
\description{
Title
}
\examples{
ideo <- create_ideo(chr="chr4",ideo.width=400)
ideo <- ideo_add_highlight(ideo, region=c(50000,60000))


}
