% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{pairend-class}
\alias{pairend-class}
\title{Title Pairend fashion interaction plot}
\value{
pairend object
}
\description{
Title Pairend fashion interaction plot
}
\section{Slots}{

\describe{
\item{\code{geom_pair}}{list. geom layer list.}

\item{\code{geneleft}}{gene_anno.  gene_anno. gene annotation object for the left gene.}

\item{\code{generight}}{gene_anno. gene_anno. gene annotation object for the right gene.}

\item{\code{.TopLeft_x}}{numeric.  top left coordinates of the ggplot object. Here top left also indicates the left coord of the gene left}

\item{\code{.TopRight_x}}{numeric. top right coordinates of the ggplot object. Here top right also indicates the right coord of the gene right.}

\item{\code{.BotLeft_x}}{numeric.  bottom left coordinates of the ggplot object. Here botleft also indicates the leftmost coord of the reads mapping to leftgene.}

\item{\code{.BotRight_x}}{numeric. bottom right coordinates of the ggplot object. Here botright also indicates the rightmost coord of the reads mapping to rightgene.}

\item{\code{.Top_y}}{numeric. top coordinates of the ggplot object.}

\item{\code{.Bot_y}}{numeric. bottom coordinates of the ggplot object.}

\item{\code{.gr1_right}}{numeric. rightmost coordinate of geneleft.}

\item{\code{.gr2_left}}{numeric. leftmost coordinate of generight.}

\item{\code{.VEXON}}{numeric. vertical height of exon.}
}}

\examples{
Null

}
