% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairend_plot.R
\name{pairend_inter}
\alias{pairend_inter}
\title{Title pair-end inter skeleton}
\usage{
pairend_inter(GENE1_anno, GENE2_anno, R1, R2, GENE1_COLOR = "#deb210",
  GENE2_COLOR = "#668ed1", xdrift = 0, ydrift = 0, VEXON = 10,
  genename1 = "", genename2 = "")
}
\arguments{
\item{GENE1_anno}{Dataframe, a bed file format dataframe for gene1. each row represents an exon with chr, start, end.}

\item{GENE2_anno}{Dataframe, a bed file format dataframe for gene2. each row represents an exon with chr, start, end.}

\item{R1}{Dataframe, a bed file format dataframe for Read1 (first end) bed file.}

\item{R2}{Dataframe, a bed file format dataframe for Read2 (second end) bed file.}

\item{GENE1_COLOR}{String, with default "#deb210".}

\item{GENE2_COLOR}{string, with default "#668ed1".}

\item{xdrift}{numeric. x axis drift from 0.}

\item{ydrift}{numeric. y axis drift from 0.}

\item{VEXON}{numeric. verticle height of exon.}
}
\value{
pair-end inter skeleton, ggplot object
}
\description{
Title pair-end inter skeleton
}
\examples{
data(GENE1_anno,GENE2_anno,R1,R2)
pairend_skeleton <- pairend_inter(GENE1_anno,GENE2_anno,R1,R2)
ggplot()+pairend_skeleton@geom_pair
}
