% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{para-class}
\alias{para-class}
\title{Title Parallel fashion interaction plot
A plot skeleton. Two interacted genes lie in up and bottom in a parallel fashion,
where interaction identified from pairend sequence will be presented as polygon connected regions.}
\value{
para obj.
}
\description{
Title Parallel fashion interaction plot
A plot skeleton. Two interacted genes lie in up and bottom in a parallel fashion,
where interaction identified from pairend sequence will be presented as polygon connected regions.
}
\section{Slots}{

\describe{
\item{\code{geom_para}}{list. geom layer list.}

\item{\code{genetop}}{gene_anno. gene annotation object for the top gene.}

\item{\code{genebot}}{gene_anno. gene annotation object for the bottom gene.}

\item{\code{.TopLeft_x}}{numeric. top left coordinates of the ggplot object. Here top left also indicates the left coord of the gene top.}

\item{\code{.TopRight_x}}{numeric. top right coordinates of the ggplot object. Here top right also indicates the right coord of the gene top.}

\item{\code{.BotLeft_x}}{numeric. bottom left coordinates of the ggplot object. Here botleft also indicates the left coord of the bottom gene.}

\item{\code{.BotRight_x}}{numeric. bottom right coordinates of the ggplot object. Here botright also indicates the right coord of the bottom gene.}

\item{\code{.Top_y}}{numeric. top coordinates of the ggplot object.}

\item{\code{.Bot_y}}{numeric. bottom coordinates of the ggplot object.}

\item{\code{.Topcolor}}{character. top gene color.}

\item{\code{.Botcolor}}{character. bottom gene color.}

\item{\code{.VEXON}}{numeric. exon veriticle height.}

\item{\code{.VGAP}}{numeric. verticle distance between genetop and gene bottom.}
}}

\examples{
Null
}
