% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_plot.R
\name{parallel_inter}
\alias{parallel_inter}
\title{Title plot parallel interaction plot skeleton}
\usage{
parallel_inter(GENE1_anno, GENE2_anno, R1, R2, GENE1_COLOR = "#deb210",
  GENE2_COLOR = "#668ed1", xdrift = 0, ydrift = 0, genename1 = "",
  genename2 = "")
}
\arguments{
\item{GENE1_anno}{Dataframe, gene annotaion bed file. eg could be seen \code{data(GENE1_anno)}}

\item{GENE2_anno}{Dataframe, gene annotaion bed file. eg could be seen \code{data(GENE2_anno)}}

\item{R1}{Dataframe, Read1 annotation file. eg could be loaded with \code{data(R1)}}

\item{R2}{Dataframe, Read2 annotation file. eg could be loaded with \code{data(R2)}}

\item{GENE1_COLOR}{String, with default "#deb210"}

\item{GENE2_COLOR}{string, with default "#668ed1"}

\item{xdrift}{numeric. x axis drift from 0.}

\item{ydrift}{numeric. y axis drift from 0.}
}
\value{
para object
}
\description{
Title plot parallel interaction plot skeleton
}
\examples{
data(GENE1_anno,GENE2_anno,R1,R2)
para <- parallel_inter(GENE1_anno,GENE2_anno,R1,R2)
ggplot() + para@geom_para
}
