% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_plot.R
\name{parallel_plot}
\alias{parallel_plot}
\title{Title Parallel interaction plot
One-station function for parallel interaction plot}
\usage{
parallel_plot(GENE1_anno, GENE2_anno, R1, R2, GENE1_COLOR = "#deb210",
  GENE2_COLOR = "#668ed1", genome = "hg38", genename1 = "",
  genename2 = "")
}
\arguments{
\item{GENE1_anno}{Character, path to gene1 annotation bed file. Gene1 will be shown at bottom}

\item{GENE2_anno}{character, path to gene2 annotation bed file. Gene2 will be shown at top.}

\item{R1}{character, path to Read1 (first end ) bed file}

\item{R2}{character, path to Read2 (second end) bed file}

\item{GENE1_COLOR}{character, In R, colors can be specified either by name e.g col = 'red' or as a hexadecimal RGB triplet}

\item{GENE2_COLOR}{character.In R, colors can be specified either by name e.g col = 'red' or as a hexadecimal RGB triplet}
}
\value{
ggplot object
}
\description{
Title Parallel interaction plot
One-station function for parallel interaction plot
}
\examples{
data(GENE1_anno,GENE2_anno,R1,R2)
para <- parallel_plot(GENE1_anno,GENE2_anno,R1,R2)

}
