% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R2}
\alias{R2}
\title{READ2 bed file dataframe}
\format{A data frame with 25 rows and 6 variables:
\describe{
  \item{v1}{chromosome name (could either be NC^ or Chr^)}
  \item{V2}{reads start}
  \item{V3}{reads end}
  \item{V4}{reads id}
  \item{V5}{score}
  \item{V6}{strand}
}}
\usage{
R2
}
\description{
A dataset containing the read1(end1) information in bed format:
chr, start, end, reads_name, score, strand. Only \strong{first 4 columns} will be used later.
column4 reads id should be corporated with R2 file reads name, the same name indicates a pair relationship.
}
\keyword{datasets}
