% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairend_plot.R
\name{pairend_inter}
\alias{pairend_inter}
\title{Title pair-end inter skeleton}
\usage{
pairend_inter(GENE1_anno, GENE2_anno, R1, R2, GENE1_COLOR = "#deb210",
  GENE2_COLOR = "#668ed1", xdrift = 0, ydrift = 0, VEXON = 10)
}
\arguments{
\item{GENE1_anno}{Character, path to gene1 annotation bed file. Gene1 will be shown at left.eg could be seen \code{data(GENE1_anno)}.}

\item{GENE2_anno}{character, path to gene2 annotation bed file. Gene2 will be shown at right. eg could be seen \code{data(GENE2_anno)}.}

\item{R1}{Dataframe, Read1 annotation file. eg could be loaded with \code{data(R1)}.}

\item{R2}{Dataframe, Read2 annotation file. eg could be loaded with \code{data(R2)}.}

\item{GENE1_COLOR}{String, with default "#deb210".}

\item{GENE2_COLOR}{string, with default "#668ed1".}

\item{xdrift}{numeric. x axis drift from 0.}

\item{ydrift}{numeric. y axis drift from 0.}

\item{VEXON}{numeric. verticle height of exon.}
}
\value{
pair-end inter skeleton, ggplot object
}
\description{
Title pair-end inter skeleton
}
\examples{
data(GENE1_anno,GENE2_anno,R1,R2)
pairend_skeleton <- pairend_inter(GENE1_anno,GENE2_anno,R1,R2)
ggplot()+pairend_skeleton@geom_pair
}
