% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairend_plot.R
\name{pairend_plot}
\alias{pairend_plot}
\title{Title Pairend interaction plot}
\usage{
pairend_plot(GENE1_anno, GENE2_anno, R1, R2, GENE1_COLOR = "#deb210",
  GENE2_COLOR = "#668ed1", genename1 = "", genename2 = "",
  xdrift = 0, ydrift = 0, VEXON = 10, genome = "hg38")
}
\arguments{
\item{GENE1_anno}{Dataframe, a bed file format dataframe for gene1. each row represents an exon with chr, start, end.}

\item{GENE2_anno}{Dataframe, a bed file format dataframe for gene2. each row represents an exon with chr, start, end.}

\item{R1}{Dataframe, a bed file format dataframe for Read1 (first end) bed file.}

\item{R2}{Dataframe, a bed file format dataframe for Read2 (second end) bed file.}

\item{GENE1_COLOR}{character, color of left gene. In R, colors can be specified either by name e.g col = "red" or as a hexadecimal RGB triplet}

\item{GENE2_COLOR}{character, color of right gene. In R, colors can be specified either by name e.g col = "red" or as a hexadecimal RGB triplet}

\item{xdrift}{numeric. x axis drift from 0.}

\item{ydrift}{numeric. y axis drift from 0.}

\item{VEXON}{numeric. verticle height of exon.}

\item{genome}{String. Genome version eg."hg38","mm10","mm20". default "hg38"}
}
\value{
ggplot object of pairend interaction plot
}
\description{
Title Pairend interaction plot
}
\examples{
data(GENE1_anno,GENE2_anno,R1,R2)
pairend <- pairend_plot(GENE1_anno,GENE2_anno,R1,R2)

}
